+++ deep_speech_2/infer.py
@@ -4,25 +4,16 @@ from __future__ import division
 from __future__ import print_function
 
 import argparse
-import distutils.util
+import functools
 import paddle.v2 as paddle
 from data_utils.data import DataGenerator
 from model import DeepSpeech2Model
 from error_rate import wer, cer
+from utils import add_arguments, print_arguments
 
-
-def add_arg(argname, type, default, help, **kwargs):
-    type = distutils.util.strtobool if type == bool else type
-    parser.add_argument(
-        "--" + argname,
-        default=default,
-        type=type,
-        help=help + ' Default: %(default)s.',
-        **kwargs)
-
-
-# yapf: disable
 parser = argparse.ArgumentParser(description=__doc__)
+add_arg = functools.partial(add_arguments, argparser=parser)
+# yapf: disable
 add_arg('num_samples',      int,    10,     "# of samples to infer.")
 add_arg('trainer_count',    int,    8,      "# of Trainers (CPUs or GPUs).")
 add_arg('beam_size',        int,    500,    "Beam search width.")
@@ -65,8 +56,8 @@ add_arg('specgram_type',    str,
         'linear',
         "Audio feature type. Options: linear, mfcc.",
         choices=['linear', 'mfcc'])
-args = parser.parse_args()
 # yapf: disable
+args = parser.parse_args()
 
 
 def infer():
@@ -114,13 +105,6 @@ def infer():
               (target, result))
         print("Current error rate [%s] = %f" %
               (args.error_rate_type, error_rate_func(target, result)))
-
-
-def print_arguments(args):
-    print("-----------  Configuration Arguments -----------")
-    for arg, value in sorted(vars(args).iteritems()):
-        print("%s: %s" % (arg, value))
-    print("------------------------------------------------")
 
 
 def main():
