+++ deep_speech_2/tools/compute_mean_std.py
@@ -4,24 +4,16 @@ from __future__ import division
 from __future__ import print_function
 
 import argparse
+import functools
 import _init_paths
 from data_utils.normalizer import FeatureNormalizer
 from data_utils.augmentor.augmentation import AugmentationPipeline
 from data_utils.featurizer.audio_featurizer import AudioFeaturizer
+from utils import add_arguments, print_arguments
 
-
-def add_arg(argname, type, default, help, **kwargs):
-    type = distutils.util.strtobool if type == bool else type
-    parser.add_argument(
-        "--" + argname,
-        default=default,
-        type=type,
-        help=help + ' Default: %(default)s.',
-        **kwargs)
-
-
-# yapf: disable
 parser = argparse.ArgumentParser(description=__doc__)
+add_arg = functools.partial(add_arguments, argparser=parser)
+# yapf: disable
 add_arg('num_samples',      int,    2000,    "# of samples to for statistics.")
 add_arg('specgram_type',    str,
         'linear',
@@ -33,15 +25,8 @@ add_arg('manifest_path',    str,
 add_arg('output_path',    str,
         'mean_std.npz',
         "Filepath of write mean and stddev to (.npz).")
-args = parser.parse_args()
 # yapf: disable
-
-
-def print_arguments(args):
-    print("-----------  Configuration Arguments -----------")
-    for arg, value in sorted(vars(args).iteritems()):
-        print("%s: %s" % (arg, value))
-    print("------------------------------------------------")
+args = parser.parse_args()
 
 
 def main():
